samples/catdogapi/requirements.txt
requests
pillow
fastapi
uvicorn
